--- src/components/old/BreadcrumbItem.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const BreadcrumbItem = ({ disabled, onClick, className, children }) => (
-  <div className={`${className}`} onClick={onClick} disabled={disabled}>
-    {children}
-  </div>
-);
-
-export default BreadcrumbItem;
